requirements.txt
# Requirements for DFS Education
# Install: pip install -r requirements.txt

Django==6.0.2
Pillow==12.1.1
python-dotenv==1.2.1
mysqlclient==2.2.7
